requirements.txt
elasticsearch==9.3.0
elasticsearch_dsl==8.15.4
openai==2.21.0
pandas==3.0.1
python-dotenv==1.2.1
streamlit==1.54.0
tqdm==4.67.3
